Refactor schedule registry slightly
