+++ atmo/clusters/management/commands/delete_clusters.py
@@ -2,7 +2,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, you can obtain one at http://mozilla.org/MPL/2.0/.
 from django.core.management.base import BaseCommand
-from atmo.clusters.jobs import delete_clusters
+from ..jobs import delete_clusters
 
 
 class Command(BaseCommand):
