+++ atmo/jobs/jobs.py
@@ -2,7 +2,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, you can obtain one at http://mozilla.org/MPL/2.0/.
 import newrelic.agent
-from atmo.jobs.models import SparkJob
+from .models import SparkJob
 
 
 @newrelic.agent.background_task(group='RQ')
